vendor: add the required SDPA mask dtype cast (4th part of the float32-leak fix)
fast-min crashed with 'Mask type must promote to output type float16': once the dtype
fixes make q/k/v fp16, the float32 additive attention mask aborts SDPA. The mask cast is
a REQUIRED companion, not optional polish — and a Metal-free unit test can't catch it
(no SDPA), only a real-model run. Recompiles clean; delta tests still 6/6.

+++ SpeechHelper/Sources/SpeechEngine/VoxtralRealtimeDecoder.swift
@@ -133,7 +133,11 @@ final class VoxtralRealtimeDecoderAttention: Module {
             let causal = kPos .<= qPos
             let window = kPos .>= (qPos - MLXArray(Int32(slidingWindow - 1)))
             let allowed = logicalAnd(causal, window)
-            let mask = MLX.where(allowed, MLXArray(0.0), MLXArray(-1e9))
+            // LOCAL FIX (see VENDORED.md): build the additive mask in the activation dtype.
+            // Once the hidden state is fp16 (the dtype fixes above), a float32 mask makes
+            // scaled_dot_product_attention abort ("Mask type must promote to output type
+            // float16"). Required companion to the float32-leak fix, not optional polish.
+            let mask = MLX.where(allowed, MLXArray(0.0), MLXArray(-1e9)).asType(q.dtype)
             maskMode = .array(mask)
         }
